[setup.py] Fix Pypi install (#29)
Fixes #26: correct Pypi install through proper handling of both flat and hierarchical file layouts.

Reviewed-on: https://codeberg.org/tfardet/mpl_chord_diagram/pulls/29
Co-committed-by: Tanguy Fardet <tfardet@noreply.codeberg.org>

+++ setup.py
@@ -3,18 +3,10 @@ from setuptools import setup, find_packages
 from shutil import rmtree, copy2
 
 
-# create directory
+# create mpl_chord_diagram directory and move important files
 setup_dir = os.path.abspath(os.path.dirname(__file__))
 directory = os.path.join(setup_dir, 'mpl_chord_diagram')
 
-try:
-    os.makedirs(directory)
-except OSError as e:
-    if e.errno != errno.EEXIST:
-        raise
-
-
-# move important files
 move = (
     '__init__.py',
     '_info.py',
@@ -26,9 +18,29 @@ move = (
 )
 
 
-for fname in move:
-    copy2(fname, directory)
+# Layout changes between git repo (flat) and Pypi (hierarchical) so check if
+# flat layout files exist. If not, then all files are already in the subfolder
 
+dir_preexists = False
+
+if not os.path.exists(directory):
+    os.makedirs(directory)
+
+    for fname in move:
+        copy2(fname, directory)
+else:
+    dir_preexists = True
+
+
+# Get version information, recommended here
+# https://packaging.python.org/guides/single-sourcing-package-version
+info = {}
+
+with open(os.path.join(directory, "_info.py")) as fp:
+    exec(fp.read(), info)
+
+
+# Package information and setup instructions
 long_descr = '''
 The module enables to plot chord diagrams with matplotlib from lists,
 numpy or scipy sparse matrices.
@@ -36,42 +48,39 @@ It provides tunable parameters such as arc sorting based on size or distance,
 and supports gradients to better visualize source and target regions.
 '''
 
-# Recommended here https://packaging.python.org/guides/single-sourcing-package-version/#single-sourcing-the-package-version
-info = {}
-with open("_info.py") as fp:
-    exec(fp.read(), info)
+try:
+    setup(
+        name='mpl_chord_diagram',
+        version=info['__version__'],
+        description='Python module to plot chord diagrams with matplotlib.',
+        package_dir={'': '.'},
+        packages=find_packages('.'),
 
-# install
-setup(
-    name='mpl_chord_diagram',
-    version=info['__version__'],
-    description='Python module to plot chord diagrams with matplotlib.',
-    package_dir={'': '.'},
-    packages=find_packages('.'),
+        # Include the non python files:
+        package_data={'': ['LICENSE', '*.md']},
 
-    # Include the non python files:
-    package_data={'': ['LICENSE', '*.md']},
+        # Requirements
+        install_requires=['numpy', 'scipy', 'matplotlib'],
 
-    # Requirements
-    install_requires=['numpy', 'scipy', 'matplotlib'],
-
-    # Metadata
-    url='https://codeberg.org/tfardet/mpl_chord_diagram',
-    author='Tanguy Fardet',
-    author_email='tanguy.fardet@tuebingen.mpg.de',
-    license='MIT',
-    keywords='chord diagram matplotlib plotting',
-    long_description=long_descr,
-    classifiers=[
-        'Development Status :: 5 - Production/Stable',
-        'Intended Audience :: Science/Research',
-        'License :: OSI Approved :: MIT License',
-        'Natural Language :: English',
-        'Programming Language :: Python :: 3',
-        'Operating System :: OS Independent',
-        'Topic :: Scientific/Engineering :: Visualization',
-        'Framework :: Matplotlib',
-    ]
-)
-
-rmtree(directory, ignore_errors=True)
+        # Metadata
+        url='https://codeberg.org/tfardet/mpl_chord_diagram',
+        author='Tanguy Fardet',
+        author_email='tanguy.fardet@tuebingen.mpg.de',
+        license='MIT',
+        keywords='chord diagram matplotlib plotting',
+        long_description=long_descr,
+        classifiers=[
+            'Development Status :: 5 - Production/Stable',
+            'Intended Audience :: Science/Research',
+            'License :: OSI Approved :: MIT License',
+            'Natural Language :: English',
+            'Programming Language :: Python :: 3',
+            'Operating System :: OS Independent',
+            'Topic :: Scientific/Engineering :: Visualization',
+            'Framework :: Matplotlib',
+        ]
+    )
+finally:
+    # only useful for package generation from git repo
+    if not dir_preexists:
+        rmtree(directory, ignore_errors=True)
